Extract terminal-tab region into sections/tab-drag.js
Hoists the 'Tab drag' region (937 LOC, covering pointer-event drag,
tab create/close/rename, split/resize, shortcuts, restore-on-reload)
out of dashboard-main.js into a new sections/tab-drag.js module.

Shared state that crossed the boundary (_termTabs, _termActive,
_termLastFocused) is replaced by a single exported termState object
so both modules can mutate it without running into ESM's read-only-
import-binding rule. _termNextId is module-local to tab-drag.js.

Three non-terminal startup calls that happened to live in the region
(checkUpdates, _startEventStream, detectSetup IIFE) stay in
dashboard-main.js; terminal side-effect registration now runs through
an exported initTabDragSideEffects() called at the bottom of
dashboard-main.js, keeping the circular import with tab-drag.js safe.

First of four region-split PRs (P-07 of P-07..P-10). No behaviour
change; see conception/projects/2026-04/2026-04-23-condash-frontend-
extraction/notes/01-p07-tab-drag-split.md for the design.
